started working on getting the team stats page to render with correct data
