+++ Scouting/Scouting.iml
@@ -130,7 +130,6 @@
     <orderEntry type="library" name="com.google.android.gms:play-services-drive-8.4.0" level="project" />
     <orderEntry type="library" name="com.android.support:support-annotations:23.1.1@jar" level="project" />
     <orderEntry type="library" name="com.squareup.okio:okio:1.13.0@jar" level="project" />
-    <orderEntry type="library" name="__local_aars__:C.\Users\Merge\Documents\GitHub\MCMergeManager\Scouting\libs\Jama-1.0.3.jar:unspecified@jar" level="project" />
     <orderEntry type="library" name="com.android.support:recyclerview-v7-23.1.1" level="project" />
     <orderEntry type="library" name="com.google.android.gms:play-services-basement-8.4.0" level="project" />
   </component>
