+++ Scouting/src/main/AndroidManifest.xml
@@ -106,6 +106,10 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="ca.team2706.scouting.mcmergemanager.gui.MainActivity" />
         </activity>
+        <activity android:name=".powerup2018.gui.TeamStatsActivity"
+            android:screenOrientation="portrait">
+
+        </activity>
     </application>
 
 </manifest>
