+++ Scouting/src/main/java/ca/team2706/scouting/mcmergemanager/powerup2018/dataObjects/TeleopScoutingObject.java
@@ -17,7 +17,7 @@ public class TeleopScoutingObject implements Serializable {
 
     /**
      * Puts all events during teleop mode into a list, sorted by timestamp.
-     * Useful for doing cycle analysis in StatsEngine2018.
+     * Useful for doing cycle analysis in StatsEngine.
      */
     public ArrayList<Event> getEvents() {
         Collections.sort(events);
